profile-image-service/requirements.txt
flask
zeep
python-dotenv
flask-cors
boto3